removing ollama segment to push to main
removing ollama due to main environment pushing into local instance without an llm available, will reintroduce later
